Make local installation easier with a script
For those who prefer to use Homebrew instead of Nix, rather than make
them perform a bunch of steps manually, we can automate the entire
setup with a single command.

Here's how the script works:

1. There's a trussworks/fresh-brew repo that has a main script that
only installs the bare minimum any dev would need: Homebrew (which also
installs Apple's Command Line Tools if they're not there yet), and Git.
I also threw in GitHub's CLI tool since it's a nice thing to have. This
script is idempotent, meaning it can be run over and over safely. If you
already have Homebrew, Git and the GH CLI, it will automatically update
them.
2. The script also allows you to customize it by adding more tools you
want to install, or commands you want to run, on a project by project
basis. To install additional tools that you want to install via
Homebrew, you add a `Brewfile.local` to your project's repo, like the
one I added here that installs `yarn` and `nodenv`. Additional
commands and bash scripts go in a `fresh-brew.local` file, like the
one I added in this PR which uses `nodenv` to install the Node version
specified in the local `.node-version` file. This means that if we
want to update the Node version for everyone, all we need to do is
update the `.node-version` file, and people can run the `fresh-press`
script from the `fresh-brew` repo and it will update everything
automatically.

## Notes for reviewers
To test this, if you already have `nodenv` and/or `yarn` installed,
please uninstall them. Then, pull this branch to your machine, and run
the command in the README:

```
bash <(curl -s https://raw.githubusercontent.com/trussworks/fresh-brew/main/fresh-press)
```

This single command should take care of the entire setup, and should
automatically run the server and launch the docs repo in your browser.

Signed-off-by: Moncef Belyamani <moncef@truss.works>