Add support for golang alias in code blocks
Currently, Docusaurus, by way of PrismJS, only supports the `go` alias
for syntax highlighting code blocks. For example, this will result in
the code being highlighted:

```
```go
```

but this will not:

```
```golang
```

In my case, I use Sublime Text, and the Markdown plugin I use only
highlights the code in my editor using the `golang` alias. We also have
many instances of the `golang` alias in the code blocks in this repo
because that's what we used in the GitHub Wiki.

PrismJS allows adding aliases to existing languages, but that would
require submitting a PR to the Prism repo, and I'm not sure how long it
would take for them to merge it.

Alternatively, Docusaurus allows you to define additional languages,
which is easy, and is what I did in this PR. Per the instructions, I ran

```
yarn run swizzle @docusaurus/theme-classic prism-include-languages
```
which created the `prism-include-languages.js` file, to which I added
`require('/src/components/prism-golang.js');`, and then I created the
`prism-golang.js` file, and copied and pasted the Go definition from the
Prism repo, and replace `go` with `golang`.

References:
https://docusaurus.io/docs/markdown-features/code-blocks#supported-languages
https://prismjs.com/extending.html#creating-a-new-language-definition
https://github.com/PrismJS/prism/blob/master/components/prism-go.js

To test this, visit https://transcom.github.io/mymove-docs/docs/dev/contributing/backend/API-Errors
and go all the way at the bottom. Notice how the last code block is not
highlighted. Now, run this PR locally, and open that same file, and
notice how it is highlighted now.