Move Vale Vocab around;
The added benefit of this commit is that it can be used to verify that
the GitHub Action supports new Multi Vocab syntax. Version 2.13.0 does
not currently support this feature, but the `errata-ai/vale` codebase
has this support. The Go binary must be built from source to get this
working going forward.