Let's make the call-out for external docs instead;
I originally referenced the external docs that I wanted folks to read
but realized that it could change and then my docs would be out of date.
This admonition is my attempt at standardizing two things. The first
thing is informing the reader that they've encountered documentation
that _should not_ be written internally. Reading the source
documentation is usually the best source of truth and there's more
incentive for the original documentation to keep things up to date. Also
upstream documentation is sometimes open source and helps level up
people to contribute and collaborate. The second thing is that we should
can use admonitions more in general to call-out specific things
regardless of the reason.