Adding a hideable sidebar;
The documentation says this is helpful when wanting the reader to focus
on the documentation. I like it because I have vertical monitors and
appreciate the option to hide the sidebar so I can better read the
content. It also works well to read the documentation on a tablet as
well.